Rename single_point_decimal
